refactor other communicators
