+++ chainermn/communicators/dummy_communicator.py
@@ -12,7 +12,7 @@ class DummyCommunicator(_base.CommunicatorBase):
     """
 
     def __init__(self, mpi_comm):
-        super(DummyCommunicator, self).__init__(mpi_comm, use_nccl=True)
+        super(DummyCommunicator, self).__init__(mpi_comm)
 
         self.gpu_buffer_a = _memory_utility.DeviceMemory()
 
@@ -20,8 +20,6 @@ class DummyCommunicator(_base.CommunicatorBase):
         _communication_utility.broadcast_naive(self.mpi_comm, model)
 
     def allreduce_grad(self, model):
-        self._init_comms()
-
         params = _memory_utility.extract_params(model)
         itemsize = 4
         n_elems_total = sum(param.grad.size for param in params)
