+++ chainermn/communicators/flat_communicator.py
@@ -8,7 +8,7 @@ from chainermn.communicators import _memory_utility
 class FlatCommunicator(_base.CommunicatorBase):
 
     def __init__(self, mpi_comm):
-        super(FlatCommunicator, self).__init__(mpi_comm, False)
+        super(FlatCommunicator, self).__init__(mpi_comm)
 
         self.gpu_buffer_a = _memory_utility.DeviceMemory()
         self.gpu_buffer_b = _memory_utility.DeviceMemory()
@@ -17,8 +17,6 @@ class FlatCommunicator(_base.CommunicatorBase):
         _communication_utility.broadcast_naive(self.mpi_comm, model)
 
     def allreduce_grad(self, model):
-        self._init_comms()
-
         params = _memory_utility.extract_params(model)
         itemsize = 4
         n_elems_total = sum(param.grad.size for param in params)
